backend/requirement.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
mysqlclient
python-dotenv
mysql-connector-python
